[Ref] ITDecompression: Make use of BitReader class. [Mod] OpenMPT: Version is now 1.28.00.19
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@10264 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/SampleIO.cpp
@@ -186,12 +186,12 @@ size_t SampleIO::ReadSample(ModSample &sample, FileReader &file) const
 					uint8 hibyte;
 					if(is16bit)
 					{
-						lowbyte = chunk.ReadBits(8);
+						lowbyte = static_cast<uint8>(chunk.ReadBits(8));
 					}
 					bool sign = chunk.ReadBits(1) != 0;
 					if(chunk.ReadBits(1))
 					{
-						hibyte = chunk.ReadBits(3);
+						hibyte = static_cast<uint8>(chunk.ReadBits(3));
 					} else
 					{
 						hibyte = 8;
@@ -199,7 +199,7 @@ size_t SampleIO::ReadSample(ModSample &sample, FileReader &file) const
 						{
 							hibyte += 0x10;
 						}
-						hibyte += chunk.ReadBits(4);
+						hibyte += static_cast<uint8>(chunk.ReadBits(4));
 					}
 					if(sign)
 					{
